refactor(code-generation): put all code_generation langs in langs/
Similarly to how it was done for treesitter a while back.
