--- lua/refactoring/code_generation/javascript.lua
@@ -1,3 +1,0 @@
--- javascript = typescript
-local typescript = require("refactoring.code_generation.typescript")
-return typescript
