+++ lua/refactoring/code_generation/langs/javascript.lua
@@ -1,0 +1,3 @@
+-- javascript = typescript
+local typescript = require("refactoring.code_generation.langs.typescript")
+return typescript
